Add socket provider and put default value on views directories
